basic-setup: Drop /etc/nsswitch.conf creation
It turns out that moss triggers did not correctly respect the dependency
order specified in the triggers, making it necessary to create
/etc/nsswitch.conf manually before attempting to install systemd.

Systemd is supposed to create /etc/nsswitch.conf itself first thing if
it doesn't already exist, but due to the aforementioned bug it didn't,
which in turn made moss error out with missing user/group errors.

However, this is now fixed, so the workaround here is no longer needed.

Signed-off-by: Rune Morling <ermo@serpentos.com>